refactor(archtest): YAML-QUOTE-FUNNEL-01 迁移到 module-path-agnostic [#1302]
移到非 test .go、const 经 PlatformModulePath 派生、dogfood 单源、baseline 缩 1；行为零变（DetectsViolation/AliasBypass/LiteralBypass + ratchet GREEN）。
